+++ src/components/pages/Profile.js
@@ -1,11 +1,5 @@
-import PropTypes from 'prop-types';
 import { useSelector } from 'react-redux';
 import { selectAllMissions } from '../../redux/missions/missionsSlice';
-
-const MissionItem = ({ missionName }) => <li>{missionName}</li>;
-MissionItem.propTypes = {
-  missionName: PropTypes.string.isRequired,
-};
 
 const Profile = () => {
   const missions = useSelector(selectAllMissions);
@@ -18,10 +12,7 @@ const Profile = () => {
         <h2>My Missions</h2>
         <ul className="missionsList">
           {joinedMissions.map((mission) => (
-            <MissionItem
-              key={mission.mission_id}
-              missionName={mission.mission_name}
-            />
+            <li key={mission.mission_id}>{mission.mission_name}</li>
           ))}
         </ul>
       </section>
@@ -29,10 +20,7 @@ const Profile = () => {
         <h2>My Rockets</h2>
         <ul className="missionsList">
           {reservedRockets.map((rocket) => (
-            <MissionItem
-              key={rocket.id}
-              missionName={rocket.name}
-            />
+            <li key={rocket.id}>{rocket.name}</li>
           ))}
         </ul>
       </section>
